Added abstract
Beer pong is a drinking game played typically on college campuses across the United States. Beer pong and its derivatives has simple rules where by arrangements of Solo® cups filled with alcoholic beverage are arranged on a table and players take turns shooting a ping pong ball into one of the cups. Use of multiple plastic cups and inconsiderate party guests litters the party location. This places a financial and psychological burden on the party host due to post-game cleanup. Here we present an alternative to beer pong that is just as, if not more, challenging. We propose the next generation of inebriated entertainment by designing a robotic cup that moves and detects successful shots. We use a previously described line following robot to follow a track. This robot features an infrared LED reflectance array and implements PID control to accurately follow the line created using a reflective tape such as electrical tape. We design a new feature to the line following algorithm for stopping on the line at certain predefined locations. The robot features a basket on top that accepts shots and detects successful shots using a photoresistor. The robot will then wait for a successful shot into the basket. The robot will then advance to the next predefined location. The winner is the user who takes the fewest shots to advance the robot to all of the predefined locations. Although one game is described, many game modes can be programmed.